pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "public_transport_gtfs"
version = "1.0.0"
description = "MacBoard query adapter for GTFS-based public transport schedules"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "httpx>=0.27",
    "pyyaml>=6.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
